# lib/api-zod/src/generated/types/createSleepLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSleepLogBody {
  /** Date of the log (YYYY-MM-DD) */
  logDate: Date;
  /** Bed time in minutes from midnight */
  bedtimeMinutes: number;
  /** When they tried to fall asleep (mins from midnight) */
  sleepAttemptMinutes: number;
  /** 1-5 scale (1=Wired/Anxious, 5=Very Relaxed) */
  eveningMood: number;
  eveningNotes?: string | null;
}