# lib/api-zod/src/generated/types/morningCheckInBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MorningCheckInBody {
  /** Last wake time in minutes from midnight */
  finalWakeTimeMinutes: number;
  /** When they got out of bed in minutes from midnight */
  outOfBedMinutes: number;
  /** How long to fall asleep in minutes (7, 10, 22, 45, 75) */
  sleepLatencyMinutes: number;
  /** 0=No, 1=Once, 2=2-3 times, 3=More than 3 times */
  wakeCount: number;
  /** Total awake time during night in minutes (5, 20, 45, 75) */
  wakeDurationMinutes?: number;
  /** 1-5 (1=Very Poor, 5=Very Good) */
  sleepQuality: number;
  /** 1-5 (1=Not at all, 5=Very Well Rested) */
  restfulness: number;
}