# lib/api-zod/src/generated/types/nightCompletion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NightCompletionChecklistItemsItem } from "./nightCompletionChecklistItemsItem";

export interface NightCompletion {
  id: number;
  userId: string;
  nightNumber: number;
  checklistItems: NightCompletionChecklistItemsItem[];
  completed: boolean;
  completedAt?: Date | null;
  createdAt: Date;
}