# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailySleepStat } from "./dailySleepStat";

export interface ProgressSummary {
  currentSleepScore?: number | null;
  firstSleepScore?: number | null;
  logsCount: number;
  currentStreak: number;
  nightsCompleted: number;
  dailyStats: DailySleepStat[];
  avgSleepLatencyMinutes?: number | null;
  avgTotalSleepMinutes?: number | null;
  avgSleepEfficiencyPct?: number | null;
  avgWakeCount?: number | null;
  avgSleepQuality?: number | null;
  insights: string[];
}