# lib/api-zod/src/generated/types/sleepLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SleepLog {
  id: number;
  userId: string;
  logDate: Date;
  bedtimeMinutes?: number | null;
  sleepAttemptMinutes?: number | null;
  eveningMood?: number | null;
  eveningNotes?: string | null;
  finalWakeTimeMinutes?: number | null;
  outOfBedMinutes?: number | null;
  sleepLatencyMinutes?: number | null;
  wakeCount?: number | null;
  wakeDurationMinutes?: number | null;
  sleepQuality?: number | null;
  restfulness?: number | null;
  /** Calculated TIB */
  timeInBedMinutes?: number | null;
  /** Calculated TST */
  totalSleepMinutes?: number | null;
  /** Calculated SE (0-100) */
  sleepEfficiencyPct?: number | null;
  /** Overall score 0-100 */
  sleepScore?: number | null;
  morningComplete: boolean;
  createdAt: Date;
}