# lib/api-zod/src/generated/types/updateNightCompletionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateNightCompletionBodyChecklistItemsItem } from "./updateNightCompletionBodyChecklistItemsItem";

export interface UpdateNightCompletionBody {
  checklistItems: UpdateNightCompletionBodyChecklistItemsItem[];
  completed?: boolean;
}