# lib/api-zod/src/generated/types/updateSleepProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSleepProfileBodySleepDisruptorFrequency } from "./updateSleepProfileBodySleepDisruptorFrequency";
import type { UpdateSleepProfileBodySleepDisruptorPrimary } from "./updateSleepProfileBodySleepDisruptorPrimary";

export interface UpdateSleepProfileBody {
  sleepDisruptorPrimary?: UpdateSleepProfileBodySleepDisruptorPrimary;
  sleepDisruptorFrequency?: UpdateSleepProfileBodySleepDisruptorFrequency;
  /** Minutes from midnight (e.g. 22:30 = 22*60+30=1350) */
  usualBedtimeMinutes?: number;
  /** Minutes from midnight */
  neededWakeUpMinutes?: number;
  /** e.g. [melatonin, herbal_supplements, sleep_apps, alcohol, nothing, other] */
  triedSolutions?: string[];
  /** Derived type like The Overthinker, The Party Recoverer, etc. */
  sleepProfileType?: string;
  /** Evening reminder time in minutes from midnight */
  reminderNightMinutes?: number;
  /** Morning reminder time in minutes from midnight */
  reminderMorningMinutes?: number;
}