# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  name?: string | null;
  sleepDisruptorPrimary?: string | null;
  sleepDisruptorFrequency?: string | null;
  usualBedtimeMinutes?: number | null;
  neededWakeUpMinutes?: number | null;
  triedSolutions?: string[] | null;
  sleepProfileType?: string | null;
  onboardingComplete: boolean;
  reminderNightMinutes?: number | null;
  reminderMorningMinutes?: number | null;
  /** Current night number (1-7) */
  currentNight: number;
  createdAt: Date;
}